Refactor to Homespun Local Storage Proxy
There's no good reason to use amplify. It's heavy, it's a bower
dependency, and it's pretty clear how to replace it with our own
solution.

Going forward the ideal way to access storage is through the "storage"
service. Please see its implementation for details. A `BrowserStore` lib
is provided for backward compatibility.

Speaking of backward compatibility, we can read and migrate old amplify
keys and values.

TECH-491